+++ app/scripts/actions/sessions.js
@@ -1,11 +1,15 @@
-export const CREATE_SESSION = 'CREATE_SESSION';
+export const CREATE_SESSION = 'server/CREATE_SESSION';
 export const DESTROY_SESSION = 'DESTROY_SESSION';
 export const FETCH_SESSIONS = 'server/FETCH_SESSIONS';
 export const SESSIONS_RECEIVED = 'SESSIONS_RECEIVED';
 
-export const createSession = () => ({
-  type: CREATE_SESSION,
-});
+const redirect = id => location.replace(`${id}/dashboard`);
+
+export const createSession = () => dispatch =>
+  dispatch({
+    type: CREATE_SESSION,
+    data: redirect,
+  });
 
 export const destroySession = id => ({
   type: DESTROY_SESSION,
