+++ lib/db.js
@@ -14,7 +14,7 @@ const getTitle = room =>
 const fetch = (room) => {
   const payload = { room };
 
-  return pool.query(`SELECT * FROM ${room} ORDER BY ID ASC`)
+  return pool.query(`SELECT * FROM ${room} ORDER BY ID DESC`)
     .then((res) => {
       payload.data = res.rows;
       return room !== 'sessions' ? getTitle(room) : Promise.resolve;
@@ -42,7 +42,7 @@ const getId = () =>
     .catch(err => error(err));
 
 const create = () =>
-  pool.query("INSERT INTO sessions(title, created) VALUES ('', CURRENT_TIMESTAMP)")
+  pool.query("INSERT INTO sessions(title, created) VALUES ('untitled session', CURRENT_TIMESTAMP)")
     .then(getId)
     .catch(err => error(err));
