+++ lib/socket.js
@@ -30,9 +30,12 @@ const open = (server) => {
       switch (action.type) {
         case 'server/FETCH_SESSIONS':
           return db.fetch('sessions').then((payload) => {
-            console.log(payload);
             io.emit('action', { type: 'SESSIONS_RECEIVED', payload });
           });
+
+        case 'server/CREATE_SESSION':
+          return db.create().then(id => action.data(id));
+
         default:
           return null;
       }
