+++ package.json
@@ -27,12 +27,12 @@
     "babel-preset-react": "^6.23.0",
     "css-loader": "^0.27.3",
     "ejs": "^2.5.6",
-    "eslint": "^3.17.1",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-loader": "^1.6.3",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-loader": "^1.7.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
     "express": "^4.15.2",
     "extract-text-webpack-plugin": "^2.1.0",
     "import-glob-loader": "^1.1.0",
